requirements.txt
pandas>=2.2.0,<3.0.0
numpy>=1.26.0,<2.0.0
scikit-learn>=1.4.0,<2.0.0
prophet>=1.1.4
joblib==1.3.2
pyyaml==6.0.1
xgboost==2.0.0
statsmodels>=0.14.1
seaborn>=0.13.0
plotly>=5.18.0
matplotlib>=3.8.0